Resource file reading and networks import
The import_networks playbook and role, and import_network module are
for importing networks specifically.

The read_resources module is meant for all importing in general. It
loads the resource YAML file structure into an Ansible variable so
that individual resources can be iterated on inside the roles.

Small deserialization utility functions are added too, for
transforming serialized data into parameters for OpenStack SDK when
creating/updating resources.

There is a block in functional tests that isn't very nice but it can
be fully removed once we implement name-based filtering on export:

https://github.com/os-migrate/os-migrate/issues/22