Export idempotence common code, tests, idempotence for Networks
This adds export idempotence common code - when replacing a resource
in a resource YAML file, we first check if indeed anything changed. If
it didn't, we don't overwrite that resource and don't even write into
the file.

This is now utilized for Network resources specifically. Some things
coming from OpenStack API seem to have non-deterministic order,
e.g. the list of networks itself, or list of subnet IDs in a
Network. To make the idempotence work properly, i've added sorting
where applicable.

Functional tests which re-execute the export and check that the file
content is the same are added.