Network import idempotence
For idempotence of imports we largely have all the pieces, all that
needs to be done is to serialize the existing network into our format
and compare if the resulting 'params' are the same as the ones
provided to the import_network module.

Crucial is looking at 'params' only and not 'info', as 'params'
contain the important network parameters that are meant to be
migrated, while 'info' contains non-migratable info, e.g. the
network's UUID.