Fix qos_policy ref in Network and remove TODOs
Fix a case of copypasta in network_refs_from_ser function.

Remove the TODOs related to the Network refs. Originally i wanted to
load also project_name and subnet_names list, but those are not
necessary for importing under current OS-Migrate assumptions, and
fetching this information in network_refs_from_sdk and
network_refs_from_ser would slow down the import/export with
unnecessary API requests. For now let's assume this info does not need
to be queried, we can re-evaluatethe decision if we get any hints that
this info would be useful.