Use tags in functional tests instead of separate playbooks
To run subset of functional tests, the current pattern was to create a
full playbook which would only include the desired parts.

An alternative pattern might be to use --skip-tags or --tags with the
test_all.yml playbook. This might allow better mix-and-match execution
for particular test environments and developer scenarios.

Due to env vars not being automatically passed into the toolbox
container, currently there are two ways to test this approach e.g. by
skipping the network tests:

    ./toolbox/shell
    FUNC_TEST_ARGS='--skip-tags test_network' make test-func

and

    ./toolbox/run bash -c "FUNC_TEST_ARGS='--skip-tags test_network' make test-func"

The 2nd case perhaps deserves some improvement (list of env vars we
want to detect and pass into toolbox container automatically?) but
that's beyond the scope of this commit.