Decide resource sameness on export by ID rather than name
This change is fueled by work on validations. We have essentially 2
options there:

* Export all resources including ones with empty name and duplicate
  names. Then run validatiton on the exported files. This requires
  that resources are iterated over for export by ID and not by name.

  This is nicer because the files can be perhaps edited by hand to
  amend the naming on destination cloud without forcing name changes
  on source cloud. Validation can be re-ran arbitrarily on the files
  to iterate forward.

* Validate the state in the source cloud to verify whether export can
  be ran, and change the state in the source cloud until unique naming
  is achieved. Then run export.

  This is less nice because amendments of the source cloud are
  necessary before an export can be performed.

To achieve the former (nicer) option, we need to start exporting by
ID. First step, which is done in this commit, is to amend the resource
sameness check which is used during export to figure out if a resource
should be added into a file or if it's already present.