Allow duplicate names on export - iterate over resources by ID
We want to export the resources as they are, even if they'd fail
validation e.g. due to duplicate naming. Validation will be performed
on the exported files and/or prior to importing.

This requires an amendment to stringfilter to be able to filter a list
of dicts and not just a simple list of strings. The `attribute`
optional argument is added. This is consistent with behavior of
Ansible's `sort` filter. By default it operates on flat lists, but
with `attribute` argument it can be made to operate on lists of dicts.