Initialize vagrant storage pool in vagrant-destroy if needed
This way you can call ./vagrant-destroy right after starting
vagrant-shell. Previously it would complain that the pool doesn't
exist. The libvirt session for vagrant is transient and the pool gets
lost every time the container is stopped, but the storage content is
not lost and the pool can simply be recreated on top of the same dir.