Assert correct os_migrate_version when loading files
Currently we do not guarantee any backward/forward compatibility in
serialized files. Using different version of OS-Migrate between export
and import isn't supported. Should users get to a situation where
differing versions are needed between export and import, they will
have to investigate the risks and e.g. `sed` the version in files to
the expected one. We're defaulting to being defensive here.

I was thinking whether this could go into validations as one of the
validation errors, but in fact if the versions are different, we
cannot even guarantee that the validations will work correctly, so
let's be careful and just raise a clear exception right away.

Closes https://github.com/os-migrate/os-migrate/issues/20