Functional CI fixes
This moves the container build earlier in the functional CI. There
seem to be flakes recently in the image build, and moving it earlier
in the CI steps should help us fail faster.

Also it makes sure that devstack is ran from /home/stack when it is
ran as stack user. We hit weird ownership issues when running as
`stack` user but devstack is under `/home/runner`.