Idea: Some OOP refactoring
The purpose of this patch was to demonstrate an area I think OOP
could be saving us time in creating resources.  I focused solely
on the export for now and only the non-nested resources.  I thought
it prudent to just do enough of a patch to get the point across, but
not too much in case we just throw it away.

IMO the main savings is reducing the amount of boilerplate to create
an export.  The SubnetResource and NetworkResource classes illustrate
how a little bit of OOP can help us just focus on adding enough code
to describe the types, info and parameters and not need to add the rest.