Introduce Resource class
We created a Resource class as the base class for all our exported
resources. It improves the DRYness of our overall code and is suitable
for both export and import. It uses method composition internally so
that it can be easily "disconnected from network" for purposes of unit
testing.

The common interfaces are now backwards compatible, e.g. the
add_or_replace method for exporting resources now accepts both direct
resource data or an instance of Resource.

Co-Authored-By: Ryan Brady <rbrady@redhat.com>