Router interfaces import
Import interfaces which Routers have in tenant Subnets. This is a more
advanced resource than the basic "CRUD" approach. First we need to
create a port with the desired IP addresses in the desired subnets,
and then we attach those ports to the router. This ensures that the
router interface IP addresses in tenant networks are migrated without
change.