Remove Transible code to cleanup the repo structure
We're not using it, and where OS-Migrate is heading now, it doesn't
seem we will literally need to take parts from it. It can only confuse
folks who try to understand OS-Migrate, making them think that the
code under openstack_ansible directory is being used for the
migration. If we need to revive any of that code, we can find it in
git history.