Fixes Subnet Update
Intermittently, the subnet import functional tests fail for ip_version
or network_id attributes.  This patch adds a readonly_params collection
to the Resource class and strips out any readonly properties during an
attempt to update.  For the Subnet resource this patch removes the
blacklisted properties in neutron source for the subnet resource
(project_id, network_id) from the update args.

A follow up issue and series of patches is forthcoming for adding
functional testing for the update on all resources.

closes #142