Update dev docs
The dev docs for running functional tests went out of date as we
changed auth mechanism and as we changed how to run a subset of
functional tests. They are now updated to current workflows.

Also slight edit has been made to the design doc now that we've made
some implementation decisions where previously there was uncertainty.