Adds test for Router Interface idempotence
This patch adds a test to ensure router interface idempotence and
updates the router interface resource class to handle the case
when a router interface already exists when attempting to create
an identical one.

This patch also fixes a small bug in the router interface where
it was using the "subnet_name" key where it should have been
using a "subnet_id" key.  This was causing the unintended behavior
of never finding a matching port when checking for ports that may
have already been created, so it always tried to create the port
leading to conflicts.