Fix subnet update test
The lineinfile module only makes sure one line is present. So after it
does its replacement in one spot, it stops replacing. Due to random
order of subnets in the file, it sometimes replaces the description of
the wrong subnet, resulting in intermittent failures. To make the
tests work, we simply replace descriptions of both subnets in the func
test, and then we check only one.

To keep tests consistent, use 'replace' instead of 'lineinfile' module
in other tests too.