Prefix and shorten export variable names
We had a bunch of unprefixed variables like:

export_networks_name_filter

We should add the os_migrate prefix and we can perhaps shorten the
variable name to still remain descriptive enough, like this:

os_migrate_networks_filter

The rename in this sense is applied across the project to all affected
variables.

This rename also opens up the possibility to use this filter on
imports in the future, allowing to work on full YAML files while
selecting just a few resources to import. It cannot be done right now
because the stringfilter module only supports looking at direct
attribute and not nested ones, so we cannot look at ['params']['name']
presently. An enhancement for stringfilter could be implemented.