Add _migration dict into resource serialization
Currently we have two dicts within the resource serialization:

* params - parameters relevant for resource creation in destination
  cloud, and for idempotence,

* _info - additional information about the resource from source cloud,
  which will not be passed on into the destination cloud when creating
  the resource.

With work on migration of rebuildable instances, we will add a third one:

* _migration - parameters for the migration process itself, not
  capturing properties of the resource, but rather *how* we want the
  resource to be migrated.

The example use case here is the workload migration. There is one
server in the source cloud but we have at least two options how to
move it to destination:

* full copy (current method) - root disk is copied as a volume, and
  the instance in the destination cloud is booted from that volume,

* rebuild (to be implemented) - instance is booted from the same
  glance image in the destination cloud, putting the destination
  instance into a similar state as after using `openstack server
  rebuild` on the original. This is better suited for instances which
  are stateless or only save state on external volumes, not on the
  root disk.

To distinguish how we want a given server migrated, we will have a
parameter in the _migration section.