Update packages when deploying conversion hosts
I ran into a bug when trying to migrate workloads, the volume export
step crashed with:

qemu-img: Unable to initialize gcrypt

It turns out there was desync in package versions, we were installing
latest qemu-img but libgcrypt was old. We need to update to latest
when deploying the conversion hosts.

Resolves: https://github.com/os-migrate/os-migrate/issues/322