Chg: Resilient RHEL unregistration mechanism
Previously the deletion of RHEL conversion hosts would require the
hosts to exist and always execute the unregisteration from RHSM. But
if at least one of the hosts didn't exist, or the host got somehow
damaged and wasn't reachable via ssh, the playbook would fail and
would not delete the conversion hosts. The deletion playbook is now
resilient in these scenarios and progresses towards ensuring the
conversion hosts are deleted.