Dev: Mechanism for skipping falsey params
Empty lists usually don't make sense to feed into the API requests as
parameters. Often the results of feeding an empty list vs. not
specifying the parameter can be the same, but it can matter in some
cases, e.g. if Neutron router availability zone extension is disabled,
the parameter must not be specified in the request.