Chg: Don't feed availability_zone_hints into requests when empty
The availability_zone_hints parameter for routers and networks can
often be an empty list. Previously we've fed it as-is into the API
requests. However, if the extension for network/router availability
zones isn't enabled, this can cause the request to fail.

We now only use the availability_zone_hints parameter when its not
empty. This should have no effect on correctness, but it can prevent
false-alarm failures.