Fix: Loosen OpenStack operation timeouts in workload migration
The current timeout for OpenStack operations during workload migration
is 10 minutes. This can be too short when working with large storage
volumes. We should make the timeout configurable, but as first
alleviation we increase the timeout to 30 minutes. This should still
be enough to fail in reasonable time when things get stuck, but it
shouldn't produce as many false alarms.