Dev: Add after-update (and after-create) hook mechanism to resources
Since it is reasonable to expect that similar or the same actions will
take place in after create and after update, we have a single
`_hook_after_update` method which gets called in both events. It has
an `is_create` parameter in case distinguishing between the two events
is necessary.