Dev: Automatic OS_MIGRATE_* variable passing through ./toolbox/run
This will make it less confusing to pass variables to make targets, as
long as they are prefixed with 'OS_MIGRATE_'. E.g. instead of starting
a nested bash shell like this:

    ./toolbox/run bash -c "E2E_TEST_ARGS='<some special args>' make test-e2e-tenant"

We will be able to set the args for the outer toolbox script and they
get automatically passed through, e.g.:

    export OS_MIGRATE_E2E_TEST_ARGS='<some special args>'
    ./toolbox/run make test-e2e-tenant