Dev: Allow virtualenv requirements override for testing
If OS_MIGRATE_REQUIREMENTS_OVERRIDE is set when we run tests, pip in
virtualenv (typically inside a toolbox container) will install that
requirements file prior to running the tests, upgrading or downgrading
dependencies as needed. This is useful to test with older openstacksdk
for example, as the toolbox image defaults to latest.

When running tests using the toolbox container as documented,
virtualenv is already pre-populated in the container image, so the
overrides only last for the life of the single container and do not
affect the test environment persistently.