Dev: Introduce lower bound requirements file
The new venv-requirements-lower-bound.txt is meant for runtime
dependencies only. In testing, it is meant to be layered on top of the
basic requirements file via the OS_MIGRATE_REQUIREMENTS_OVERRIDE
mechanism.

We start with the same requirements as in the main requirements file
which will allow us to start using the override in CI without risk of
breaking. Then we can edit the lower bound in another pull request and
fix any related issues.