Fix: Compatibility with OpenStack SDK 0.36.x
It seems like the only place we're hitting problems is using 'owner'
property of an image resource. Both 0.36 and 0.52 SDK support the
'owner_id' property there, so let's just use that.

Interestingly enough, this is not the case for the query parameters in
`conn.image.images()`. There `image_id` is not working anywhere, it
has to be `image` in both 0.36 and 0.52, so we'll stick to that.