Dev: Run shell when venv-run is passed no arguments
In toolbox dir, we have the run/shell script pairs, one for running a
command in a container, another for launching a shell in a
container. However, it can be done so that if no arguments are passed
to the 'run' script, it acts the same way as the 'shell' script. This
is already true for the plain 'run' and the 'vagrant-run', only
'venv-run' is edited to achieve the behavior described above.