Fix: Skip migration failure cleanup when there is nothing to clean up
Previously the source/destination cleanup would be attempted on a
failed migration, even though the migration didn't progress far enough
to create the variables that are needed for cleanup. This would then
produce a false alarm error during the cleanup. Now it's fixed, the
cleanups are attempted only if there is necessary info available (if
the migration progressed far enough for the cleanup to be necessary).