New: Import of users' Nova keypairs as admin
Import of Nova keypairs can now be done for other users as well,
provided that the authenticated user running the playbook is
authorized to read and create keys for other users (an admin
user is typically authorized to do that). The playbook to perform this
import is called `import_users_keypairs.yml`.

Both `os_migrate_users_filter` and `os_migrate_keypairs_filter`
variables are respected when iterating over users and their keypairs.