Fix: Allow exporting same-named keypairs for different users
Keypair IDs are typically the same as their names, they aren't
randomized UUIDs. We cannot rely on them for distinguishing
uniqueneess on export and performing idempotency checks when exporting
to a file which already contains some resources. The sameness check
now uses ID + user_ref tuple, making it possible to export keys with
the same ID values for mutliple users.