Dev: Add minimal runtime.yml for the Ansible collection
Publishing 0.9.0 failed with:

ERROR! Galaxy import process failed: 'requires_ansible' in
meta/runtime.yml is mandatory, but no meta/runtime.yml found (Code:
None)

It seems there is a new requirement for Ansible collections of adding
meta/runtime.yml [1], and the minimal runtime.yml seems to consist of
supported range of Ansible versions. Let's start defensively with 2.9
only, as that's what we currently CI, and expand/move the list
according to what we test.

We need to exclude this file from linting, as ansible-lint in Ansible
2.9 is not yet working with runtime.yml and fails on:

701 No 'galaxy_info' found
/root/os_migrate/os_migrate/meta/runtime.yml:1
{'meta/main.yml': {'requires_ansible': '>=2.9,<2.10', '__line__': 1, '__file__': '/root/os_migrate/os_migrate/meta/runtime.yml'}}

[1] https://github.com/ansible/ansible/commit/a9eb8b04882669bd17cd82780c908863e7504886