Fix: Filter read only properties from SDK on import rather than export
This will work better than filtering on import, because whatever is in 'params' plays
a role in idempotency checking. So if we're not using 'self' or 'stores' when creating
the resource, it should not be in 'params' at all, as it could be causing
"false alarms" when checking whether a resource needs updating in destination
cloud or not.