Dev: Remove description from volume_map
Description is now not used in the volume_map in workload
migration. Description of volumes is transferred/set via the `volumes`
section of server's `params`. In case a new boot volume is created
that didn't exist in the source cloud, boot volume description can be
set via `boot_volume_params` migration paramter.