Dev: Only advise running build when releasing
The procedure in releasing docs is very unlikely to break tests, we
don't need to advise running them manually before posting a release
PR. In the unlikely case that the release changes do break tests, the
CI will catch the problem.

However, we need to advise running a build at least, as the const.py
file is updated automatically during a build.