Dev: Fix ansible-lint by uninstalling Ansible RPM
There is an issue with ansible-lint - it fails with "ImportError:
cannot import name 'AnsibleCollectionLoader' from
'ansible.utils.collection_loader'" when Ansible is installed on the
host and we're running lint in a venv.

We're working around it by uninstalling Ansible in the `test-lint`
make target. We're checking that we run in a toolbox container before
doing the uninstall.

https://github.com/ansible-community/ansible-lint-action/issues/41#issuecomment-933663572

Co-Authored-By: Adrian Brown <adrbrown@redhat.com>